+++ src/components/property-page-widgets/Amenities/component.js
@@ -7,6 +7,7 @@ import { GridColumn } from 'layout/GridColumn';
 import { Heading } from 'typography/Heading';
 import { Link } from 'elements/Link';
 import { Modal } from 'elements/Modal';
+import { VIEW_MORE } from 'utils/default-strings';
 
 import { getDefaultItems } from './utils/getDefaultItems';
 import { hasExtraItems } from './utils/hasExtraItems';
@@ -16,7 +17,12 @@ import { getCategoryMarkup } from './utils/getCategoryMarkup';
  * The standard widget for displaying the amenities of a property.
  * @returns {Object}
  */
-export const Component = ({ amenities, headingText, isStacked }) => (
+export const Component = ({
+  amenities,
+  headingText,
+  isStacked,
+  modalTriggerText,
+}) => (
   <Grid stackable>
     {headingText && (
       <GridColumn width={12}>
@@ -28,7 +34,7 @@ export const Component = ({ amenities, headingText, isStacked }) => (
     )}
     {hasExtraItems(amenities, isStacked) && (
       <GridColumn width={12}>
-        <Modal trigger={<Link>View more</Link>}>
+        <Modal trigger={<Link>{modalTriggerText}</Link>}>
           <SemanticModal.Content>
             <Grid padded stackable>
               {amenities.map(getCategoryMarkup)}
@@ -45,6 +51,7 @@ Component.displayName = 'Amenities';
 Component.defaultProps = {
   headingText: null,
   isStacked: false,
+  modalTriggerText: VIEW_MORE,
 };
 
 Component.propTypes = {
@@ -65,4 +72,6 @@ Component.propTypes = {
   headingText: PropTypes.string,
   /** Are the amenities displayed stacked on top of one another */
   isStacked: PropTypes.bool,
+  /** The text for the modal trigger */
+  modalTriggerText: PropTypes.string,
 };