+++ src/components/property-page-widgets/Amenities/utils/getFormattedAmenityItems.js
@@ -1,3 +1,5 @@
+import { AND } from 'utils/default-strings';
+
 /**
  * @param  {String[]} items
  * @return {String[]}
@@ -6,7 +8,7 @@ export const getFormattedAmenityItems = items =>
   items.map((item, index) => {
     // Is penultimate item in array
     if (index === items.length - 2) {
-      return `${item} and `;
+      return `${item} ${AND} `;
     }
     // Is last item in array
     if (index === items.length - 1) {