+++ styleguide/config/options/template.js
@@ -2,4 +2,5 @@ const { TITLE } = require('../constants');
 
 module.exports = {
   title: TITLE,
+  favicon: './favicon.ico',
 };